ChangeList: The date and time arguments to _dos_getftime() and _dos_setftime() are now using 'unsigned int' type instead of 'unsigned short'. This change has been made to improve compatibility with other compilers.
Note: Tell 32-bit users that they need to recompile. Suggest
using stat()/utime() instead of non-portable functions.

[git-p4: depot-paths = "//depot/openwatcom/": change = 31998]

+++ bld/clib/_dos/c/sftimwnt.c
@@ -24,8 +24,7 @@
 *
 *  ========================================================================
 *
-* Description:  WHEN YOU FIGURE OUT WHAT THIS FILE DOES, PLEASE
-*               DESCRIBE IT HERE!
+* Description:  Implementation of _dos_setftime() for Win32.
 *
 ****************************************************************************/
 
@@ -37,11 +36,12 @@
 #include "ntex.h"
 #include "seterrno.h"
 
-_WCRTLINK unsigned _dos_setftime( int hid, WORD date, WORD time )
+
+_WCRTLINK unsigned _dos_setftime( int hid, unsigned date, unsigned time )
 {
     int         error;
     HANDLE      h;
-    FILETIME    ctime,atime,wtime;
+    FILETIME    ctime, atime, wtime;
 
     error = 0;
     h = __getOSHandle( hid );
